backend/requirements.txt
Django==3.1
psycopg2-binary==2.8.6
djangorestframework==3.12
django-cors-headers==3.7.0